// client/src/generated/client/models/RsShortLobbyInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Play with friends
 * Play with friends
 *
 * The version of the OpenAPI document: 1.0
 * Contact: theo.dm94@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    RsParticipant,
    RsParticipantFromJSON,
    RsParticipantFromJSONTyped,
    RsParticipantToJSON,
} from './RsParticipant';

/**
 * 
 * @export
 * @interface RsShortLobbyInfo
 */
export interface RsShortLobbyInfo {
    /**
     * 
     * @type {string}
     * @memberof RsShortLobbyInfo
     */
    lobbyID: string;
    /**
     * 
     * @type {Array<RsParticipant>}
     * @memberof RsShortLobbyInfo
     */
    players: Array<RsParticipant>;
    /**
     * 
     * @type {boolean}
     * @memberof RsShortLobbyInfo
     */
    isInGame: boolean;
}

export function RsShortLobbyInfoFromJSON(json: any): RsShortLobbyInfo {
    return RsShortLobbyInfoFromJSONTyped(json, false);
}

export function RsShortLobbyInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): RsShortLobbyInfo {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'lobbyID': json['lobbyID'],
        'players': ((json['players'] as Array<any>).map(RsParticipantFromJSON)),
        'isInGame': json['isInGame'],
    };
}

export function RsShortLobbyInfoToJSON(value?: RsShortLobbyInfo | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'lobbyID': value.lobbyID,
        'players': ((value.players as Array<any>).map(RsParticipantToJSON)),
        'isInGame': value.isInGame,
    };
}